subscription: move from `peer` to `conference`
