+++ pkg/conference/subscription/rewriter/expand.go
