+++ pkg/conference/subscription/rewriter/rewriter.go
